+++ src/theme/variables/edgeDark.ts
@@ -328,6 +328,10 @@ export const edgeDark: Theme = {
   underlayColor: palette.white,
   underlayOpacity: 0.95,
 
+  // TouchableHighlights overlay
+  overlayDisabledColor: palette.blackOp50,
+  overlayDisabledTextColor: palette.white,
+
   // Tutorials
   tutorialModalUnderlay: palette.transparent,